feat: add basketItem

+++ BakendProject/Models/BasketItem.cs
@@ -14,10 +14,11 @@ namespace BakendProject.Models
         public User User { get; set; }
         public string AppUserId { get; set; }
         public AppUser AppUser { get; set; }
-        public int Sum { get; set; }
+        public double Sum { get; set; }
         public int Count { get; set; }
         public double Price{ get; set; }
         public string ImgUrl { get; set; }
+        
 
     }
 }
